requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pyarrow==14.0.1
pandas==2.1.4
orjson==3.9.10
lxml==4.9.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
pydantic==2.5.0
